+++ jdaviz/configs/default/plugins/markers/markers.py
@@ -5,8 +5,7 @@ from traitlets import Bool, Unicode, Instance, observe, default
 import ipywidgets as widgets
 
 from jdaviz.core.events import (ViewerAddedMessage, ChangeRefDataMessage,
-                                AddDataMessage, RemoveDataMessage, MarkersPluginUpdate,
-                                GlobalDisplayUnitChanged)
+                                AddDataMessage, RemoveDataMessage, MarkersPluginUpdate)
 from jdaviz.core.marks import MarkersMark, DistanceMeasurement
 from jdaviz.core.registries import tray_registry
 from jdaviz.core.template_mixin import (PluginTemplateMixin, ViewerSelectMixin, TableMixin,
@@ -120,9 +119,6 @@ class Markers(PluginTemplateMixin, ViewerSelectMixin, TableMixin):
         self.hub.subscribe(self, RemoveDataMessage,
                            handler=lambda msg: self._recompute_mark_positions(msg.viewer))
 
-        self.hub.subscribe(self, GlobalDisplayUnitChanged,
-                           handler=self._on_units_changed)
-
         self.docs_description = """Press 'm' with the cursor over a viewer to log
                                  the mouseover information. To change the
                                  selected layer, click the layer cycler in the
@@ -133,17 +129,6 @@ class Markers(PluginTemplateMixin, ViewerSelectMixin, TableMixin):
                                  measurement points to the nearest marker."""
         # description displayed under plugin title in tray
         self._plugin_description = 'Create markers on viewers.'
-
-    def _on_units_changed(self, msg=None):
-        """
-        Callback for when the global display units are changed.
-        This clears all distance measurements to prevent them from being
-        displayed in the wrong locations.
-        """
-        if len(self._distance_marks) > 0:
-            # The clear_table method handles clearing both the UI table
-            # and the marks from all viewers.
-            self.measurements_table.clear_table()
 
     @default('measurements_table')
     def _default_measurements_table(self):
@@ -574,7 +559,7 @@ class Markers(PluginTemplateMixin, ViewerSelectMixin, TableMixin):
                 preview_text = "" if is_profile else "..."
 
                 self._temporary_dist_measure = DistanceMeasurement(
-                    viewer, start_x, start_y, start_x, start_y, text=preview_text
+                    viewer, x=[start_x, start_x], y=[start_y, start_y], text=preview_text
                 )
                 self._temporary_dist_measure.label_shadow.visible = False
                 for mark in self._temporary_dist_measure.marks:
@@ -608,9 +593,8 @@ class Markers(PluginTemplateMixin, ViewerSelectMixin, TableMixin):
                     self._viewer_for_drawing = None
                     return
 
-                dist_measure = DistanceMeasurement(viewer, plot_x0, plot_y0,
-                                                   plot_x1, plot_y1,
-                                                   text=text_plot)
+                dist_measure = DistanceMeasurement(viewer, x=[plot_x0, plot_x1],
+                                                   y=[plot_y0, plot_y1], text=text_plot)
                 dist_measure.endpoints = {'p1': p1, 'p2': p2}
                 self._distance_marks.setdefault(viewer_id, []).append(dist_measure)
                 viewer.figure.marks = viewer.figure.marks + dist_measure.marks